cards v0.1.0 Cards

Provides methods for creating and handleing a deck of cards

Functions

contains?(deck, card)

Determines whether a dack contains a given card throught the card argument

Example

iex> deck = Cards.create_deck
iex> Cards.contains?(deck, "Ace of Spades")
true

create_deck()

Creates a list of strings that represents the deck of cards

create_hand(hand_size)

deal(deck, hand_size)

Divides a deck into a hand and the rest of the cards. The hand_size argument determine how many cards will be in the hand

Examples

  iex> deck = Cards.create_deck
  iex> {hand, _deck} = Cards.deal(deck, 1)
  iex> hand
  ["Ace of Spades"]

load(filename)

save(deck, filename)

shuffle(deck)